Move all containers to the containers folder and change imports in App
