+++ src/containers/App/index.js
@@ -2,11 +2,11 @@ import React, { Component } from 'react';
 import './App.css';
 import Header from '../Header';
 import Sidebar from '../Sidebar';
-import LandingPage from '../../components/LandingPage';
-import Bio from '../../components/Bio';
-import Skills from '../../components/Skills';
+import LandingPage from '../LandingPage';
+import Bio from '../Bio';
+import Skills from '../Skills';
 import Projects from '../Projects';
-import Contact from '../../components/Contact';
+import Contact from '../Contact';
 import ProjectModal from '../ProjectModal';
 import { withRouter, Switch, Route } from 'react-router-dom';
 import { connect } from 'react-redux';
